renamed Lenses folder to Deflectors.
